Dead lock fix & event queue size fix
A non-blocking read/write call in BufferedSerial was creating a deadlock. Fixed
in this commit courtesy unlocking before returning from the function.

Event queue size reserved for the stream_cb() is doubled, i.e., 2 times the given
EVENT_QUEUE_SIZE. This seems to fit our callback alongwith its parameter which happens
to be a short.