Extending FileHandle & introducing mbed_poll
This has been an attempt to extend existing FileHandle to behave like POSIX
file descriptor for input/output resources/devices too.
This will now provide an abstract indicator/handle any arbitrary file or device
type resource. May be in future, sockets too.

In order to correctly detect availability of read/write a FileHandle, we needed
a select or poll mechanisms. We opted for poll as POSIX defines in
http://pubs.opengroup.org/onlinepubs/009695399/functions/poll.html
Currently, mbed_poll() just spins and scans filehandles looking for any events we
are interested in. In future, his spinning behaviour will be replaced with condition variables.

In retarget.cpp we have introduced an mbed_fdopen() function which is equivaelnt to
C fdopen(). It attaches a std stream to our FileHandle stream. newlib-nano somehow does
not seem to call isatty() so retarget doesn't work for device type file handles. We handle
this by checking ourselves in mbed_fdopen() if we wish to attach our stream to std stream.
We also turn off buffering by C library as our stuff will be buffered already.