Introducing a BufferedSerial
BufferedSerial is a FileHandle and is a Serial too.
It keeps the SerialBase private however lets the user extend FileHandle
by keeping it public.

It is using CircularBuffer class for having circular buffers implements poll()
mechanism for FileHandle. Currently mbed_poll() is just spinning _poll_change is
not implemented yet which is necessary for poll() to work properly. _poll_change()
will notify of the selectability.

There are some minor amendments in CircularBuffer too.