Detaching Trace library from tests
Trace libraray has been causing a massive usage of stack
for example MBED_TRACE_LINE_LENGTH is 1024 bits.
That's we are attempting to detach trace library totally from tests.
There will be a PR for application as well in response to this so as
to make business application and run_test.py also detached from
trace library by default.