ADR request Mask list bug fix
Setting channel mask to zero was buggy and hence it caused overflow
into other members of the struct.
Copy of the mask provided by the netwotk was also ridden with a bug.
copy_channel_mask() takes size of the mask list as parameter not the
size of the mask variable.