Removing dynamic memory allocation in PHY
set_next_channel() is called for every transmission. Arguably, this is the last
place we want to do a dynamic memory allocation because of memory fragmentation.
In LoRaPHY class we are now using a magic number 16 as we know that all EU
like PHY implementations use 16 channels and the other PHYs who use more than 16
channels are made to override this function. Hence we don't necessarily need to
allocate memory dynamically.

Same goes for link_ADR_req(). We know that the PHY layers which are EU like always
have a channel mask length of 16 bits representing 16 channels so we use the magic
number and override the method in the PHYs where there are more than 16 channels.