Swapping I2C pins between original and extended PAD
Looking at the board layout (SAML21 Xplained Pro) we can see that out of
three EXT pinouts only two (EXT1 and EXT3) have a set of four pins
available on a single SERCOM. For example (PB16, PB17, PB22, PB23) on
EXT3 are mapped to SERCOM5. So if we wish to add a peripheral that uses
4 pin SPI like a SX1276 Wing radio, we can only choose EXt3 or EXT1.

However, we found out that we can't use radio shield on EXT3 as the pin
PB10 used as dio2 interrupt line is also mapped to the LED.
For both SPI and I2C peripherlas, HAL chooses Default PAD. This would
result in picking up SERCOM0 for I2C pins (PA08 and PA09) which will not
work for us as we have attached the secure element to EXT3 and we need
to map these pins to SERCOM2.

One way of solving the issue is to mark PA08 and PA09 as NC on SERCOM0
by default. Another is to swap the SERCOMx mapping for the PA08 and PA09
between the Default PAD and Extended PAD.

A proper fix would probably be to physically cut PB10 and LED
connection. This would mean that we don't need to hack the pin map.