Adding ability to change/restore RX2 frequency
Usually RX2 frequency is fixed as per regional parameter specifications
but there can be exceptions. Network servers are allowed to change RX2
frequency using Mac commands or out of the band mechanisms. We have
introduces APIs in this commit which will enable the application to
modify or restore RX2 window frequency.