Adding Multicast register & multicast reception
The stack will now maintain a register with a maximum of 4 multicast
session entries. The maximum number of session contexts is dictated by
the LoRaWAN remote multicast setup specifcation v1.0.0.

'register_multicast_address()' API is used to set an entry into the
multicast register, 'get_multicast_addr_register()' API is used to acquire a
handle to the multicast register itself and
'verify_multicast_freq_and_dr()' API is used to verify frequency and
data rate parameters if requested by the network in a special case
(i.e., a switch to class C at a given time with a given frequency and SF).

When a multicast downlink is received, the stack looks-up the multicast
register for the address match. If a match is found, other session
specific parameters in the register entry, i.e., frmae ocunters and security
keys are loaded to be used for the stack for successful reception and
decryption. The application is notified about the reception as it would
normally be notified (i.e., RX_DONE event).

If the downlink doesn't match any of the multicast addresses or the
unicast address, the packet is dropped and an adrress failure is
notified to the application.