Adding ClockSync Plugin
This plug-in or application extension implements LoRaWAN Application
Layer Clock Synchronization Specification v1.0.0. It is supposed to
exist in-tree as an on-demand plugin. Applications can use the plug-in
to synchronize their clocks with network if the protocol version is
v1.0.2 or lower. For v1.0.3 and higher, use of DevTimeReq mac command is
recommended.

The plugin is completely detached from the core stack or application
logic. It just provides easy to use interface to handle Clock sync
commands. The application is reponsible for actual transmission of
uplinks at the clock sync port and evoking the plugin upon reception of
a clock sync command.